+++ src/value/value.c
@@ -16,9 +16,9 @@ GK_EXPORT
 void
 gkInitValue(GkValue * __restrict val,
             GkType               type,
-            size_t               rowc,
-            size_t               colc,
-            size_t               itemsz) {
+            uint32_t             rowc,
+            uint32_t             colc,
+            uint32_t             itemsz) {
   val->type     = type;
   val->itemSize = itemsz;
   val->colCount = GLM_MAX(colc, 1);
@@ -36,16 +36,16 @@ void
 gkInitValueAsFloat(GkValue *val, float fval) {
   val->type           = GKT_FLOAT;
   val->s32.floatValue = fval;
-  val->rowCount = 1;
-  val->colCount = 1;
-  val->val = &val->s32.floatValue;
+  val->rowCount       = 1;
+  val->colCount       = 1;
+  val->val            = &val->s32.floatValue;
 }
 
 GK_EXPORT
 void
 gkInitValueAsVec3(GkValue *val, vec3 v) {
-  val->type  = GKT_FLOAT3;
-  val->val   = realloc(val->val, sizeof(vec3));
+  val->type     = GKT_FLOAT3;
+  val->val      = realloc(val->val, sizeof(vec3));
   val->rowCount = 3;
   val->colCount = 1;
   glm_vec3_copy(v, val->val);
@@ -54,24 +54,24 @@ gkInitValueAsVec3(GkValue *val, vec3 v) {
 GK_EXPORT
 void
 gkInitValueAsVec4(GkValue *val, vec4 v) {
-  val->type  = GKT_FLOAT4;
-  val->val   = realloc(val->val, sizeof(vec4));
+  val->type = GKT_FLOAT4;
+  val->val  = realloc(val->val, sizeof(vec4));
   glm_vec4_ucopy(v, val->val);
 }
 
 GK_EXPORT
 void
 gkInitValueAsMat3(GkValue *val, mat3 m) {
-  val->type  = GKT_FLOAT3x3;
-  val->val   = realloc(val->val, sizeof(mat3));
+  val->type = GKT_FLOAT3x3;
+  val->val  = realloc(val->val, sizeof(mat3));
   glm_mat3_copy(m, val->val);
 }
 
 GK_EXPORT
 void
 gkInitValueAsMat4(GkValue *val, mat4 m) {
-  val->type  = GKT_FLOAT4x4;
-  val->val   = realloc(val->val, sizeof(mat4));
+  val->type = GKT_FLOAT4x4;
+  val->val  = realloc(val->val, sizeof(mat4));
   glm_mat4_ucopy(m, val->val);
 }
